beckham: update blobs from beckham PPW29.183-29-1

+++ recovery/root/sbin/prepdecrypt.sh
@@ -61,36 +61,46 @@ mkdir -p /vendor/lib64/hw/
 cp /s/system/lib64/android.hidl.base@1.0.so /sbin/
 cp /s/system/lib64/libicuuc.so /sbin/
 cp /s/system/lib64/libxml2.so /sbin/
+cp /s/system/lib64/libkeymaster4support.so /sbin/
 cp /s/system/lib64/libkeymaster_messages.so /sbin/
+cp /s/system/lib64/libkeymaster_portable.so /sbin/
+cp /s/system/lib64/libsoftkeymasterdevice.so /sbin/
 
 relink /v/bin/qseecomd
 
+cp /v/lib64/hw/android.hardware.boot@1.0-impl.so /vendor/lib64/hw/
+cp /v/lib64/hw/android.hardware.gatekeeper@1.0-impl-qti.so /vendor/lib64/hw/
+cp /v/lib64/hw/android.hardware.keymaster@3.0-impl-qti.so /vendor/lib64/hw/
+cp /v/lib64/hw/bootctrl.sdm660.so /vendor/lib64/hw/
+cp /v/lib64/hw/gatekeeper.sdm660.so /vendor/lib64/hw/
+cp /v/lib64/hw/keystore.sdm660.so /vendor/lib64/hw/
 cp /v/lib64/libdiag.so /vendor/lib64/
 cp /v/lib64/libdrmfs.so /vendor/lib64/
 cp /v/lib64/libdrmtime.so /vendor/lib64/
 cp /v/lib64/libGPreqcancel.so /vendor/lib64/
 cp /v/lib64/libGPreqcancel_svc.so /vendor/lib64/
-cp /v/lib64/libqdutils.so /vendor/lib64/
-cp /v/lib64/libqisl.so /vendor/lib64/
-cp /v/lib64/libqservice.so /vendor/lib64/
-cp /v/lib64/libQSEEComAPI.so /vendor/lib64/
-cp /v/lib64/librecovery_updater_msm.so /vendor/lib64/
-cp /v/lib64/librpmb.so /vendor/lib64/
-cp /v/lib64/libSecureUILib.so /vendor/lib64/
-cp /v/lib64/libssd.so /vendor/lib64/
-cp /v/lib64/libtime_genoff.so /vendor/lib64/
 cp /v/lib64/libkeymasterdeviceutils.so /vendor/lib64/
 cp /v/lib64/libkeymasterprovision.so /vendor/lib64/
 cp /v/lib64/libkeymasterutils.so /vendor/lib64/
-cp /v/lib64/hw/bootctrl.sdm660.so /vendor/lib64/hw/
-cp /v/lib64/hw/gatekeeper.sdm660.so /vendor/lib64/hw/
-cp /v/lib64/hw/keystore.sdm660.so /vendor/lib64/hw/
-cp /v/lib64/hw/android.hardware.boot@1.0-impl.so /vendor/lib64/hw/
-cp /v/lib64/hw/android.hardware.gatekeeper@1.0-impl-qti.so /vendor/lib64/hw/
-cp /v/lib64/hw/android.hardware.keymaster@3.0-impl-qti.so /vendor/lib64/hw/
+cp /v/lib64/libqdutils.so /vendor/lib64/
+cp /v/lib64/libqisl.so /vendor/lib64/
+cp /v/lib64/libQSEEComAPI.so /vendor/lib64/
+cp /v/lib64/libqservice.so /vendor/lib64/
+cp /v/lib64/libqtikeymaster4.so /vendor/lib64/
+cp /v/lib64/librecovery_updater_msm.so /vendor/lib64/
+cp /v/lib64/librpmb.so /vendor/lib64/
+cp /v/lib64/libsecureui_svcsock.so /vendor/lib64/
+cp /v/lib64/libSecureUILib.so /vendor/lib64/
+cp /v/lib64/libssd.so /vendor/lib64/
+cp /v/lib64/libtime_genoff.so /vendor/lib64/
 
-cp /v/manifest.xml /vendor/
-cp /v/compatibility_matrix.xml /vendor/
+if [ -f /v/manifest.xml ]; then
+	cp /v/manifest.xml /vendor/
+	cp /v/compatibility_matrix.xml /vendor/
+else
+	cp /v/etc/vintf/manifest.xml /vendor/
+	cp /v/etc/vintf/compatibility_matrix.xml /vendor/
+fi
 
 relink /v/bin/hw/android.hardware.boot@1.0-service
 relink /v/bin/hw/android.hardware.gatekeeper@1.0-service-qti
